requirements.txt
django
django-bootstrap3
chardet
djangorestframework
paramiko
tornado
django-jet
ansible
celery==3.1.25
celery-with-redis==3.0
django-celery==3.2.2
FeedParser
jsonfield
django-guardian